Tests/LibWeb/Text/expected/navigation/sandboxed-history-back-cannot-traverse-top.txt

location=?second
ui-current=?second
ui-buttons=true,false
